requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Login==0.6.2
Flask-WTF==1.0.1
Flask-Migrate==3.1.0
WTForms==3.0.1
Werkzeug==2.0.3
